+++ proofs/src/plonk/verifier.rs
@@ -350,7 +350,6 @@ where
         &challenges,
     );
 
-    let const_1_com = CS::constant_commitment();
     let lin_com = compute_linearization_commitment(
         expressions,
         vk,
@@ -359,7 +358,6 @@ where
         &xn,
         &splitting_factor,
         &quotient_limb_coms,
-        &const_1_com,
     );
 
     // Collect queries that are checked in the multi-open argument